+++ e-commerce-back/src/main/java/org/mohammedezzaim/ecommerceback/bean/payment/Payment.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 public class Payment {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String ref;
 
@@ -33,14 +33,13 @@ public class Payment {
     @ManyToOne(fetch = FetchType.LAZY)
     private Administrator administrator;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
-
     public String getRef() {
         return ref;
     }
